Support Double baby

+++ dsp_utils/Source/AudioHistoryBuffer.cpp
@@ -1,11 +1,11 @@
 #include "AudioHistoryBuffer.h"
 
 AudioHistoryBuffer::AudioHistoryBuffer(int historySize)
-    : buffer(historySize, 0.0f), writePos(0), size(historySize)
+    : buffer(historySize, 0.0), writePos(0), size(historySize)
 {
 }
 
-void AudioHistoryBuffer::pushSamples(const float* samples, int numSamples)
+void AudioHistoryBuffer::pushSamples(const double* samples, int numSamples)
 {
     for (int i = 0; i < numSamples; ++i)
     {
@@ -14,7 +14,7 @@ void AudioHistoryBuffer::pushSamples(const float* samples, int numSamples)
     }
 }
 
-void AudioHistoryBuffer::getHistory(float* outSamples, int numSamples) const
+void AudioHistoryBuffer::getHistory(double* outSamples, int numSamples) const
 {
     jassert(numSamples <= size);
 
@@ -22,7 +22,7 @@ void AudioHistoryBuffer::getHistory(float* outSamples, int numSamples) const
     int firstPart = juce::jmin(size - start, numSamples);
     int secondPart = numSamples - firstPart;
 
-    std::memcpy(outSamples, buffer.data() + start, firstPart * sizeof(float));
+    std::memcpy(outSamples, buffer.data() + start, firstPart * sizeof(double));
     if (secondPart > 0)
-        std::memcpy(outSamples + firstPart, buffer.data(), secondPart * sizeof(float));
+        std::memcpy(outSamples + firstPart, buffer.data(), secondPart * sizeof(double));
 }
